# Add perf benchmarks for protobufjs