Partially address #123

Results of the perf benchmark on my machine:

```
### read binary
google-protobuf             :     413.662 ops/s
ts-proto                    :   1,324.736 ops/s
protobuf-ts (speed)         :   1,461.452 ops/s
protobuf-ts (speed, bigint) :   1,475.889 ops/s
protobuf-ts (size)          :   1,250.677 ops/s
protobuf-ts (size, bigint)  :   1,255.167 ops/s
protobufjs                  :   1,732.049 ops/s
### write binary
google-protobuf             :     906.883 ops/s
ts-proto                    :   3,805.993 ops/s
protobuf-ts (speed)         :     430.632 ops/s
protobuf-ts (speed, bigint) :     448.063 ops/s
protobuf-ts (size)          :     378.682 ops/s
protobuf-ts (size, bigint)  :     392.511 ops/s
protobufjs                  :   1,539.768 ops/s
### from partial
ts-proto                    :   4,503.332 ops/s
protobuf-ts (speed)         :   1,568.577 ops/s
protobuf-ts (size)          :   1,555.881 ops/s
### read json
ts-proto                    :   3,719.366 ops/s
protobuf-ts (speed)         :     889.613 ops/s
protobuf-ts (size)          :     890.034 ops/s
protobufjs                  :   4,120.232 ops/s
### write json
ts-proto                    :  13,200.842 ops/s
protobuf-ts (speed)         :   1,865.668 ops/s
protobuf-ts (size)          :   1,862.537 ops/s
protobufjs                  :   4,199.372 ops/s
### read json string
ts-proto                    :     957.057 ops/s
protobuf-ts (speed)         :     416.284 ops/s
protobuf-ts (size)          :     421.48  ops/s
protobufjs                  :     910.256 ops/s
### write json string
ts-proto                    :   1,000.572 ops/s
protobuf-ts (speed)         :     943.338 ops/s
protobuf-ts (size)          :     949.784 ops/s
protobufjs                  :   1,446.891 ops/s
```