@timostamm 2021-06-25
My performance numbers are off by a factor of 100. This is a bug in the makefiles, I'll address that.